Refactor: Header Section and NavBarSection Seperation of Concerns
